% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello-world.R
\name{hello_world}
\alias{hello_world}
\title{Say hello}
\usage{
hello_world()
}
\value{
The string \code{"Hello, world!"}.
}
\description{
Return a simple greeting to demonstrate the package.
}
\examples{
hello_world()
}
